Fallback to an empty Stream in UploadedFile (#21)
It is possible for StreamFactoryInterface::createStreamFromFile to fail
and throw a RuntimeException. One such case may occur when the file
upload has failed and the tmp_name is empty.

ServerRequestCreator should not exit on such an error, and should be
able to create a list of UploadedFile instances included those with
errors. When StreamFactoryInterface::createStreamFromFile fails, use an
empty Stream instead.