The `Host` header, according to [RFC2616 section 14.23](https://tools.ietf.org/html/rfc2616#section-14.23), may contain a port-number.
The current implementation doesn't handle this, and effectively ends up producing an `UriInterface` instance that returns a host-name string with the port-number inside it.

PSR-7 actually isn't explicit on this point, but I'd have to assume that, since there's a `getPort()` method for obtaining the port-number, then `getHost()` isn't supposed to also return the port-number, just the host-name.

This PR includes a regression-test (which did fail) and a fix.